Enhance UI with new Frame and update existing Frame
Added a new Frame with a VerticalStackLayout to display a Label
with the FriendlyName and a QR code image, which includes a
TapGestureRecognizer for copy commands. Updated the existing
Frame to style the "InformationAboutThingText" label with
BoldItemTitleLabel, set its LineBreakMode to WordWrap, and
HorizontalTextAlignment to Start. Commented out the
"InformationAboutThing" label to reduce UI clutter.