Add Pop3 back method and clean up NavigationArgs.cs
Added a new `Pop3` value to the `BackMethod` enum in `BackMethod.cs`
to support popping 3 pages in the navigation stack. Updated the
`GetBackRoute` method in `NavigationArgs.cs` to handle the new
`Pop3` back method, returning the route "../../..". Also, removed
an extra space in the initialization of the `NavigationCompletionSource`
field in `NavigationArgs.cs`.