Add SignView XAML and code-behind to project
Added `SignView.xaml` and `SignView.xaml.cs` to the `NeuroAccessMaui.csproj` file. These files are located in the `UI\Pages\Contracts\ViewContract\Views` directory. The `SignView.xaml` file is set to be compiled with the `MSBuild:Compile` generator, and the `SignView.xaml.cs` file is marked as dependent upon the `SignView.xaml` file.