Refactor app startup and window management
Added a new using directive for `NeuroAccessMaui.Services.Intents` in `App.xaml.cs`. Commented out the block of code that sets the `MainPage` property within the `App` constructor when `backgroundStart` is false. Introduced a new method `CreateWindow` that overrides the base class method, which checks for existing windows and returns the first one if available, or creates a new window using `AppShell`. These changes aim to improve the app's startup process and window management.