Refactor notifications to use ObservableNotification
Updated NotificationsViewModel.cs to change the type of the
notifications collection to ObservableCollection<ObservableNotification>.
Added a using directive for NeuroAccessMaui.UI.Pages.Notifications.ObjectModel
to support the new type. Modified the OnAppearing method to wrap each
NotificationEvent in an ObservableNotification before adding it to the
notifications collection. Updated the OnNotificationEvent method to
compare Notification objects instead of ObjectId, and to add new
notifications as ObservableNotification objects. Added a try-catch
block to log exceptions during this process.