Refactor theme management in ITagProfile interface
Changed the `Theme` property to be read-write in the `ITagProfile` interface, removing the `SetTheme(AppTheme Theme)` and `SetTheme()` methods. Updated the `FromConfiguration` method in `TagProfile.cs` to set the theme directly using `ServiceRef.ThemeService.SetTheme()`. The `IncrementNrPeerReviews` method now includes a setter for the `Theme` property, streamlining theme management and enhancing property access.