Add ThemeLoaded property to IThemeService interface
Introduced a new `ThemeLoaded` property of type `TaskCompletionSource` in the `IThemeService` interface to indicate when the theme has been loaded. Implemented this property in the `ThemeService` class to track the loading status. Updated the `ApplyProviderTheme` method to signal the completion of the theme loading process by calling `TrySetResult` on `ThemeLoaded`.