Enhance Theme Loading, Registration Flow, and UI Loading States (#267)
* Update app version to 2.6.5 in project files

Updated `NeuroAccessMaui.csproj` to change the
`ApplicationDisplayVersion` and `ApplicationVersion`.
Replaced `Info.plist` with a new version, updating
`CFBundleShortVersionString` while retaining all
necessary configurations and permissions.

* Add ThemeLoaded property to IThemeService interface

Introduced a new `ThemeLoaded` property of type `TaskCompletionSource` in the `IThemeService` interface to indicate when the theme has been loaded. Implemented this property in the `ThemeService` class to track the loading status. Updated the `ApplyProviderTheme` method to signal the completion of the theme loading process by calling `TrySetResult` on `ThemeLoaded`.

* Enhance ProfilePhoto logic in ViewIdentityViewModel

Updated the `ProfilePhoto` property to prioritize photos
with filenames starting with "ProfilePhoto". If none are
found, it falls back to the first photo in the collection.
This improves the logic for selecting the profile photo.

* Enhance ShowWallet with beta feature feedback

Modified the `ShowWallet` method in `AppsViewModel.cs` to check for beta features. If beta features are unavailable, a delay is introduced with a visual indication by setting the `BetaFeaturePressed` flag. If beta features are available, the method proceeds with navigation to the `WalletPage` using `WalletNavigationArgs`. This improves user experience by providing feedback before navigation.

* Enhance theme loading feedback in MainPage

Replaced static Image in MainPage.xaml with a ConditionView that displays a UriImage or ActivityIndicator based on the ThemeLoaded property. This improves user experience by providing visual feedback during theme loading.

Introduced a new observable property, themeLoaded, in MainViewModel.cs to track the theme loading state. Updated the BannerUri property to ensure it defaults to BannerUriLight when the theme is not loaded.

Modified the OnAppearing method to wait for the theme to load before updating the UI, ensuring the correct banner image is displayed based on the current theme state.

* Enhance registration flow and loading feedback

- Updated `ShowBiometricsInfo` to set `AuthenticationMethod` to `Password` and navigate to `Finalize`.
- Commented out the assignment of `Pin` in `ContactSupportViewModel.cs`.
- Changed navigation in `ContactSupportViewModel.cs` from `Finalize` to `DefinePassword`.
- Modified conditions in `CheckAndHandleIdentityApplicationAsync` to prevent unnecessary navigation.
- Added `ConditionView` in `GetStartedView.xaml` to display an `ActivityIndicator` during loading.
- Introduced `isLoading` property in `GetStartedViewModel.cs` to track loading state.
- Updated `ScanQrCode` method to manage `IsLoading` state during QR code scanning.
- Ensured `IsLoading` is reset after error handling in `GetStartedViewModel.cs`.
- Enhanced `ExistingAccount` method to include QR code scanning functionality.