Add XML schema validation
This commit introduces XML schema validation functionality to the NeuroAccessMaui application. It includes the addition of the XML service, instantiation of the XML schema validation service, and pre-registration of schemas for branding and KYC processes. Additionally, a previously commented-out line for sending alerts has been uncommented to allow alert posting to a specified URL.